Add first pass of more general integer parameterized actions
